chore: address review comments on thumbnail extraction PR (#519)

+++ crates/core/src/view/home/mod.rs
@@ -2295,7 +2295,7 @@ impl View for Home {
                 true
             }
             Event::GoTo(location) => {
-                self.go_to_page(location as usize, rq, context);
+                self.go_to_page(location, rq, context);
                 true
             }
             Event::Chapter(dir) => {
